requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
flask-sock==0.7.0
pytest==8.3.4
python-dotenv==1.0.1
PyJWT==2.10.1
flasgger==0.9.7.1
requests==2.32.3
